# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend-for-frontend API for the LLM Wiki project. Serves wiki sources and concepts with full-text search, import, and pipeline status.",
        "title": "LLM Wiki BFF API",
        "contact": {
            "name": "LLM Wiki Team"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/concepts": {
            "get": {
                "security": [
                    {
                        "DevUserAuth": []
                    },
                    {
                        "ProjectHeader": []
                    }
                ],
                "description": "Returns published wiki concepts by default. Set include_drafts=true to include draft concepts.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "concepts"
                ],
                "summary": "List wiki concepts",
                "parameters": [
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "Include draft concepts",
                        "name": "include_drafts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ConceptsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sources": {
            "get": {
                "security": [
                    {
                        "DevUserAuth": []
                    },
                    {
                        "ProjectHeader": []
                    }
                ],
                "description": "Returns all compiled wiki sources.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sources"
                ],
                "summary": "List wiki sources",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SourcesListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gcs.WikiPage": {
            "type": "object",
            "properties": {
                "concepts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "path": {
                    "type": "string"
                },
                "quality": {
                    "type": "string"
                },
                "raw_source": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "status": {
                    "description": "\"published\" or \"draft\"",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "handler.ConceptsListResponse": {
            "type": "object",
            "properties": {
                "concepts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gcs.WikiPage"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handler.SourcesListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/gcs.WikiPage"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "DevUserAuth": {
            "description": "DEV mode user identity header. Required when dev_jwt=true.",
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        },
        "ProjectHeader": {
            "description": "Project identifier header.",
            "type": "apiKey",
            "name": "X-Project-ID",
            "in": "header"
        }
    }
}